pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pukaist-engine"
version = "0.1.0"
description = "Local-first document ingestion and analysis engine"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.23",
    "python-multipart>=0.0.7",
    "requests>=2.31",
    "httpx>=0.27",
    "pypdf>=4.3",
    "pdf2image>=1.17.0",
    "python-docx>=1.1.0",
    "PyJWT[crypto]>=2.9.0",
    "fastkml>=0.12",
    "pytesseract>=0.3.13",
    "Pillow>=11.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-cov>=4.1",
    "ruff>=0.1.0",
]
hunyuan = [
    "openai>=1.40.0",
    "transformers>=4.43.0",
    "vllm>=0.6.0",
]
embeddings = [
    "accelerate>=0.33.0",
    "bitsandbytes>=0.43.0",
    "sentencepiece>=0.2.0",
    "torch>=2.3.0",
    "transformers>=4.43.0",
    "sentence-transformers>=4.1.0",
]

# =============================================================================
# Tool Configuration
# =============================================================================

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v"

[tool.coverage.run]
source = ["src"]
branch = true
omit = [
    "src/__pycache__/*",
    "src/hunyuan_ocr.py",  # Optional heavy dependency
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
show_missing = true
